Clean up text directives implementation.
