+++ content/material/Components-Gallery/99-Directives-Spike.mdx
@@ -20,7 +20,3 @@ It is possible to nest admonitions.
 :::warning[...if useful]
 However, the use case may be somewhat limited.
 :::
-
-```
-:mdi[google-chrome]
-```
