+++ src/framework/plugins/remark-text-directives/plugin.ts
@@ -3,13 +3,30 @@ import {VFile} from "vfile";
 import {Literal, Node, Parent} from "unist";
 import {visit} from "unist-util-visit";
 import {TextDirective} from "mdast-util-directive";
-import {MdxJsxElement, TextDirectiveDeclaration} from "../shared/models";
+import {EsmImport} from "../shared/models";
 import {Optional} from "../../util/optional";
 import {Log} from "../../util/log";
 import {ensureEsmImports} from "../shared/util/mdast-util-esm-imports";
+import {MdxJsxTextElement} from "mdast-util-mdx-jsx";
+import {PhrasingContent} from "mdast";
 
 export interface TextDirectivesConfig {
   declarations: TextDirectiveDeclaration<any>[];
+}
+
+export interface TextDirectiveDeclaration<T extends TextDirectiveTransformerProps> {
+  name: string;
+  transform: TextDirectiveTransformer<T>,
+  esmImports: EsmImport[];
+}
+
+export type TextDirectiveTransformer<T extends TextDirectiveTransformerProps>
+  = (attributes: T) => Optional<MdxJsxTextElement>
+
+export interface TextDirectiveTransformerProps {
+  children: PhrasingContent[],
+  literal?: string,
+  class?: string;
 }
 
 /** @type {import('unified').Plugin<[TextDirectivesConfig], MdastRoot>} */
@@ -30,7 +47,6 @@ export default function remarkTextDirectives(config: TextDirectivesConfig): Tran
 
         parent.children[parent.children.indexOf(node)] = jsxElement.get();
         ensureEsmImports(mdast, declaration.esmImports);
-        console.log(mdast);
       });
     });
   };
@@ -40,7 +56,7 @@ function matchDeclaration(config: TextDirectivesConfig, node: TextDirective): Op
   return Optional.of(config.declarations.find(declaration => declaration.name === node.name));
 }
 
-function transform(node: TextDirective, declaration: TextDirectiveDeclaration<any>): Optional<MdxJsxElement> {
+function transform(node: TextDirective, declaration: TextDirectiveDeclaration<any>): Optional<MdxJsxTextElement> {
   const children = node.children;
   const literal = children.length === 1 && children[0].type == 'text'
     ? (children[0] as Literal).value
