+++ src/framework/plugins/shared/util/jsx-node-util.ts
@@ -1,32 +1,33 @@
 import {JsxElementSpec, MdxJsxElement} from "../models";
-import {BlockContent, DefinitionContent, PhrasingContent} from 'mdast';
-
-export enum JsxElementType {
-  FLOW = 'mdxJsxFlowElement',
-  TEXT = 'mdxJsxTextElement',
-}
-
+import {PhrasingContent} from 'mdast';
+import {MdxJsxTextElement, MdxJsxAttribute} from "mdast-util-mdx-jsx";
 
 export function jsxFlowElementFrom(spec: JsxElementSpec, children?: Node[]): MdxJsxElement {
-  return createMdxJsxElement(JsxElementType.FLOW, spec, children);
-}
-
-export function jsxTextElementFrom(spec: JsxElementSpec, children?: Node[]): MdxJsxElement {
-  return createMdxJsxElement(JsxElementType.TEXT, spec, children);
-}
-
-function createMdxJsxElement(elementType: JsxElementType, spec: JsxElementSpec, children?: Node[]): MdxJsxElement {
   return {
-    type: elementType,
+    type: 'mdxJsxFlowElement',
     name: spec.componentName,
-    attributes: spec.attributes.map(attr => {
-      return {
-        type: 'mdxJsxAttribute',
-        name: attr.name,
-        value: attr.value,
-      }
-    }),
+    attributes: mapJsxAttributes(spec),
     children: children ?? [] as unknown,
     data: {_mdxExplicitJsx: true}
   } as MdxJsxElement;
+}
+
+export function jsxTextElementFrom(spec: JsxElementSpec, children?: PhrasingContent[]): MdxJsxTextElement {
+  return {
+    type: 'mdxJsxTextElement',
+    name: spec.componentName,
+    attributes: mapJsxAttributes(spec),
+    children: children ?? [],
+    data: {_mdxExplicitJsx: true}
+  };
+}
+
+function mapJsxAttributes(spec: JsxElementSpec): MdxJsxAttribute[] {
+  return spec.attributes.map(attr => {
+    return {
+      type: 'mdxJsxAttribute',
+      name: attr.name,
+      value: attr.value,
+    }
+  })
 }
